Got Add Location working
Bootstrap dialog works fine, but needs more work for the bootstrap
alerts used in the angular version.
So far this code is really crappy ... angular doesn't map easily to react
:(

This exercise so far has been very useful to get a better understanding
of the 2 different approaches, though ...

Another point of concern is that I need to click on the big + button
twice for some reason ... maybe it's just bad state management?